refactor : GET, POST /order/image/complete 라우터 분리
