+++ server/routes/OrderCompleteImage.ts
@@ -1,7 +1,7 @@
-import express, { NextFunction, Request, Response } from "express";
+import express, { Request } from "express";
 import multer from "multer";
 
-import { findImage, postImage } from "../service/Order";
+import OrderController from "../Controllers/OrderController";
 
 const storage = multer.memoryStorage();
 const upload = multer({ storage: storage });
@@ -9,30 +9,10 @@ const upload = multer({ storage: storage });
 const router = express.Router();
 
 // GET /order/image/complete
-router.get("/", async (req: Request, res: Response, next: NextFunction) => {
-  try {
-    const query = req.query;
-    const image = findImage(query);
-    res.send(image);
-  } catch (error) {
-    console.error(error);
-    next(error);
-  }
-});
+router.get("/", OrderController.getImage);
 
 // POST /order/image/complete
-router.post("/", upload.single("uploadImage"), async (req: Request, res: Response, next: NextFunction) => {
-    try {
-      const body = req.body;
-      const documentFile = (req as MulterRequest).file;
-      await postImage(body, documentFile);
-      res.send({ msg: "done" });
-    } catch (error) {
-      console.error(error);
-      next(error);
-    }
-  }
-);
+router.post("/", upload.single("uploadImage"), OrderController.postImage);
 
 export interface MulterRequest extends Request {
   file: any;
